[Manual Tests] Domain Subdomain Templates
From the "stx_test_plan_template - 20190226.txt", the following naming
conventions are proposed:

- Test ID: <domain>_<feature>_<unique-id#>
- Test Title: test_<title>
- Tags:

Example to show the proposed format for 2 different domains, one without
subdomains, the other with one or more subdomains:

- DomainX: domainx (without subdomains)
- DomainY: domainy (with subdomains)

The "Table Of Contents" suggested to show at "StarlingX Manual Tests"
page is:

- DomainX
  - Test Requirements
  - Test Cases
- DomainY
  - Test Requirements
  - Domains

There is no intermmediate "Table Of Contents" for "DomainX" since it is a
single page, index.rst will host all "Domain X" information.

In domains with one or more subdomains and since a RST file is proposed per
subdomain, an intermmediate "Table Of Contents" is suggested to show at
"DomainY" index.rst page, based in the naming convention it will show up as:

- SubdomainOne
  - domainy_subdomainone_01
  - domainy_subdomainone_02
  - domainy_subdomainone_03
- SubdomainTwo
  - domainy_subdomaintwo_01
  - domainy_subdomaintwo_02
  - domainy_subdomaintwo_03
  - domainy_subdomaintwo_04
  - domainy_subdomaintwo_05

There is a local "Table Of Content" at the top of each RST file that hosts
test cases for easy navigation.

Depends-On: https://review.openstack.org/#/c/640180/

Co-Authored-By: Juan Pablo Gomez <juan.p.gomez@intel.com>
Co-Authored-By: Fernando Hernandez Gonzalez <fernando.hernandez.gonzalez@intel.com>

Change-Id: Id7bf131f220763f09e593147653fae68f43956e3
Signed-off-by: Juan Pablo Gomez <juan.p.gomez@intel.com>
Signed-off-by: Fernando Hernandez Gonzalez <fernando.hernandez.gonzalez@intel.com>
Signed-off-by: Cristopher Lemus <cristopher.j.lemus.contreras@intel.com
Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>