[Manual Tests] Feature Templates
Example to show the proposed format for 2 different features, one without
a domain, the other with one or more domains:

- Feature X: FeatureWithoutDomain
- Feature Y: FeatureWithDomain

The Table Of Contents suggested to show at "StarlingX Manual Tests"
page is:

- FeatureWithoutDomain
  - Test Requirements
  - Test Cases
- FeatureWithDomain
  - Test Requirements
  - Domains

"FeatureWithoutDomain" is a single page.

In Features with one or more domains and since a RST file is proposed per domain,
the Table Of Contents suggested to show at "FeatureWithDomain" index page is:

- DomainOne
  - featurewithdomain_domainone_01
  - featurewithdomain_domainone_02
  - featurewithdomain_domainone_03
- DomainTwo
  - featurewithdomain_domainone_01
  - featurewithdomain_domainone_02
  - featurewithdomain_domainone_03
  - featurewithdomain_domainone_04
  - featurewithdomain_domainone_05

There is a local Table Of Content in each RST file to list Test Cases at the
top of the file.

Depends-On: https://review.openstack.org/#/c/639775

Co-Authored-By: Juan Pablo Gomez <juan.p.gomez@intel.com>
Co-Authored-By: Fernando Hernandez Gonzalez <fernando.hernandez.gonzalez@intel.com>

Change-Id: Id7bf131f220763f09e593147653fae68f43956e3
Signed-off-by: Juan Pablo Gomez <juan.p.gomez@intel.com>
Signed-off-by: Fernando Hernandez Gonzalez <fernando.hernandez.gonzalez@intel.com>
Signed-off-by: Cristopher Lemus <cristopher.j.lemus.contreras@intel.com
Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>