Networking Interface Attach Detach subdomain
With this patch the intention is to start working on Attach Detach subdomain,
this patch includes 3 basic test cases in oder to verify the behavior for
each time that we perform a attach / detach operation on different sceranrios,
this includes different drivers and making enphasis on phisicall interfaces or
virtual environments. Mainly we are just deatacching from existing port in
order to keep the steps in a straight forward way.

Change-Id: I3b307385b9d3cdba0519d74c63961f50fb3f53c1