Networking Regression Manual sub domain
The goal for this patch is to start including all the regression test cases
that contains basic operations for networking domain. These test cases can
show the networking health in our systems. The test scope was considering any
possible scenario where it should contain the operations across the time.
Making this subdomain easy to understand and highly valuable for future exe-
cutions.

Change-Id: I9597347f4f7f9f9ce2f61be271d73cfaea49adc4